Fix trip creation and modification

+++ TravelAgent/TravelAgent/Service/AccommodationService.cs
@@ -46,7 +46,7 @@ namespace TravelAgent.Service
                     {
                         Id = reader.GetInt32(0),
                         Name = reader.GetString(1),
-                        Rating = reader.GetInt32(2),
+                        Rating = reader.GetDouble(2),
                         Image = $"{_consts.PathToAccommodationImages}/{reader.GetString(3)}",
                         Location = location
                     };
@@ -87,7 +87,7 @@ namespace TravelAgent.Service
                     {
                         Id = reader.GetInt32(0),
                         Name = reader.GetString(1),
-                        Rating = reader.GetInt32(2),
+                        Rating = reader.GetDouble(2),
                         Image = $"{_consts.PathToAccommodationImages}/{reader.GetString(3)}",
                         Location = location
                     };
